Move shared utils to new file and add function for adding import statement
